+++ src/alert/api/zh-cn.js
@@ -7,34 +7,40 @@ module.exports = [
             {
                 name: 'thyType',
                 description: '指定警告提示的类型',
-                type: `'success' | 'warning' | 'danger' | 'info' | 'primary-week' | 
-                'success-week' | 'warning-week' | 'danger-week'`,
+                type: `'success' | 'warning' | 'danger' | 'info' | 'primary-weak' | 
+                'success-weak' | 'warning-weak' | 'danger-weak'`,
                 default: 'info'
             },
             {
                 name: 'thyMessage',
-                description: '警告提示的内容',
+                description: '显示警告提示的内容',
                 type: 'string',
                 default: ''
             },
             {
                 name: 'thyIcon',
-                description: '自定义图标类型，true/false 控制是否显示图标，当为字符串时，指定图标的名称',
+                description: '显示自定义图标，可传 true/false 控制是否显示图标，或者传字符串去指定图标名称',
                 type: 'boolean | string',
                 default: ''
             },
             {
                 name: 'thyCloseable',
-                description: '默认不显示关闭警告框按钮',
+                description: '是否显示关闭警告框按钮，默认不显示',
                 type: 'boolean',
                 default: 'false'
+            },
+            {
+                name: 'operation',
+                description: '警告框自定义操作',
+                type: 'ContentChild<TemplateRef>',
+                default: ''
             }
         ]
     },
     {
         type: 'directive',
         name: 'thyAlertActionItem',
-        description: '样式指令,为警告内容的操作按钮添加样式',
+        description: '样式指令,可为警告内容添加自定义操作按钮样式',
         properties: []
     }
 ];