+++ src/button/examples/block/block.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-button-block-example',
+    templateUrl: './block.component.html'
+})
+export class ThyButtonBlockExampleComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}