+++ src/button/examples/module.ts
@@ -1,15 +1,18 @@
+import { ThyButtonModule } from 'ngx-tethys/button';
+import { ThySpaceModule } from 'ngx-tethys/space';
+
+import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+
 import { ThyButtonBasicExampleComponent } from './basic/basic.component';
-import { ThyButtonPairExampleComponent } from './pair/pair.component';
+import { ThyButtonBlockExampleComponent } from './block/block.component';
+import { ThyButtonGroupExampleComponent } from './group/group.component';
 import { ThyButtonIconExampleComponent } from './icon/icon.component';
 import { ThyButtonLinkExampleComponent } from './link/link.component';
-import { ThyButtonSizeExampleComponent } from './size/size.component';
-import { ThyButtonOutlineExampleComponent } from './outline/outline.component';
-import { ThyButtonSquareExampleComponent } from './square/square.component';
 import { ThyButtonLoadingExampleComponent } from './loading/loading.component';
-import { ThyButtonGroupExampleComponent } from './group/group.component';
-import { ThyButtonModule } from 'ngx-tethys/button';
-import { CommonModule } from '@angular/common';
+import { ThyButtonOutlineExampleComponent } from './outline/outline.component';
+import { ThyButtonPairExampleComponent } from './pair/pair.component';
+import { ThyButtonSizeExampleComponent } from './size/size.component';
 
 const COMPONENTS = [
     ThyButtonBasicExampleComponent,
@@ -18,15 +21,15 @@ const COMPONENTS = [
     ThyButtonLinkExampleComponent,
     ThyButtonSizeExampleComponent,
     ThyButtonOutlineExampleComponent,
-    ThyButtonSquareExampleComponent,
     ThyButtonLoadingExampleComponent,
-    ThyButtonGroupExampleComponent
+    ThyButtonGroupExampleComponent,
+    ThyButtonBlockExampleComponent
 ];
 
 @NgModule({
     declarations: [...COMPONENTS],
     entryComponents: [...COMPONENTS],
-    imports: [CommonModule, ThyButtonModule],
+    imports: [CommonModule, ThyButtonModule, ThySpaceModule],
     exports: [...COMPONENTS]
 })
 export class ThyButtonExamplesModule {}