+++ src/flexible-text/examples/container/container.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-flexible-text-container',
+    templateUrl: './container.component.html',
+    styleUrls: ['./container.component.scss']
+})
+export class ThyFlexibleTextContainerExampleComponent implements OnInit {
+    public text = `New platforms are providing creators with a chance
+         bypass pirate sites and platform rules, and connect directly with users`;
+
+    constructor() {}
+
+    ngOnInit() {}
+}