+++ src/select/examples/top-display/top-display.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+import { taskTypes } from '../mock-data';
+
+@Component({
+    selector: 'thy-select-top-display-example',
+    templateUrl: './top-display.component.html'
+})
+export class ThySelectTopDisplayExampleComponent implements OnInit {
+    optionData = taskTypes;
+
+    selectedTaskType: string;
+
+    ngOnInit() {
+        this.selectedTaskType = this.optionData[0].name;
+    }
+}