+++ src/slide/slide-ref.service.ts
@@ -1,8 +1,10 @@
-import { Injectable } from '@angular/core';
-import { ThyAbstractInternalOverlayRef, ThyUpperOverlayPosition, ThyAbstractOverlayRef } from 'ngx-tethys/core';
-import { ThySlideContainerComponent } from './slide-container.component';
+import { ThyAbstractInternalOverlayRef, ThyAbstractOverlayPosition, ThyAbstractOverlayRef } from 'ngx-tethys/core';
+
 import { OverlayRef } from '@angular/cdk/overlay';
-import { slideUpperOverlayOptions, ThySlideConfig } from './slide.config';
+import { Injectable } from '@angular/core';
+
+import { ThySlideContainerComponent } from './slide-container.component';
+import { slideAbstractOverlayOptions, ThySlideConfig } from './slide.config';
 
 export abstract class ThySlideRef<T, TResult = any> extends ThyAbstractOverlayRef<T, ThySlideContainerComponent, TResult> {}
 
@@ -10,10 +12,10 @@ export abstract class ThySlideRef<T, TResult = any> extends ThyAbstractOverlayRe
 export class ThyInternalSlideRef<T = any, TResult = any> extends ThyAbstractInternalOverlayRef<T, ThySlideContainerComponent, TResult>
     implements ThySlideRef<T> {
     constructor(overlayRef: OverlayRef, containerInstance: ThySlideContainerComponent, config: ThySlideConfig) {
-        super(slideUpperOverlayOptions, overlayRef, containerInstance, config);
+        super(slideAbstractOverlayOptions, overlayRef, containerInstance, config);
     }
 
-    updatePosition(position?: ThyUpperOverlayPosition): this {
+    updatePosition(position?: ThyAbstractOverlayPosition): this {
         return this.updateGlobalPosition(position);
     }
 }